Api_Class_Assignments/Thrusday_Class_Assignments/Day_03/Assignment_02/pyproject.toml
[project]
name = "assignment-02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.6.5",
    "openai-agents>=0.2.4",
    "python-decouple>=3.8",
]
